Set mfs to continuous variables

+++ stree/tests/utils.py
@@ -22,6 +22,6 @@ def load_disc_dataset(
     random_state=0, n_classes=2, n_features=3, n_samples=1500
 ):
     np.random.seed(random_state)
-    X = np.random.randint(1, 17, size=(n_samples, n_features), dtype=int)
+    X = np.random.randint(1, 17, size=(n_samples, n_features)).astype(float)
     y = np.random.randint(low=0, high=n_classes, size=(n_samples), dtype=int)
     return X, y
